Remove top level check for HLS backend
@alorence suggests in
https://github.com/orcc/orcc/issues/119#issuecomment-49842465 that
overriding  `doValidate()` will allow backends to generate networks
that contain ports. This commit mirrors the override implemented in
the TTA backend in 0b4a5403bb3b557dd20a6a9bfe8ad5c7c6ee89e5 .

This commit is meant more for discussion -- I don't know if this
`doValidate()` implementation is appropriate for the HLS backend. Only
if HLS guru's think it is should it be considered for merging
upstream.